+++ apps/docs/.gitignore
@@ -8,7 +8,7 @@
 
 # test & build
 /coverage
-/.next/
+.next/
 /out/
 /build
 *.tsbuildinfo
