+++ apps/docs/content/docs/index.mdx
@@ -25,49 +25,25 @@ The generated code is built on [Codama](https://github.com/codama-idl/codama), t
 
 ## Where to start?
 
-<div className="grid sm:grid-cols-2 gap-6 mt-8">
-  <a
+<Cards>
+  <Card 
+    title="Installation" 
+    description="Step-by-step guide to install Coda. Set up your local development environment."
     href="/docs/installation"
-    className="group relative rounded-lg border border-border bg-card p-6 transition-colors hover:bg-accent"
-  >
-    <h3 className="font-semibold text-lg mb-2">Installation</h3>
-    <p className="text-sm text-muted-foreground">
-      Step-by-step guide to install Coda. Set up your local
-      development environment.
-    </p>
-  </a>
-
-  <a
+  />
+  <Card 
+    title="Quick Start" 
+    description="Generate your first TypeScript client from an Anchor IDL. No configuration required."
     href="/docs/quick-start"
-    className="group relative rounded-lg border border-border bg-card p-6 transition-colors hover:bg-accent"
-  >
-    <h3 className="font-semibold text-lg mb-2">Quick Start</h3>
-    <p className="text-sm text-muted-foreground">
-      Generate your first TypeScript client from an Anchor IDL. No configuration
-      required.
-    </p>
-  </a>
-</div>
-
-<div className="grid sm:grid-cols-2 gap-6 mt-6">
-  <a
+  />
+  <Card 
+    title="Why Coda?" 
+    description="Learn how Coda simplifies Solana development."
     href="/docs/why-coda"
-    className="group relative rounded-lg border border-border bg-card p-6 transition-colors hover:bg-accent"
-  >
-    <h3 className="font-semibold text-lg mb-2">Why Coda?</h3>
-    <p className="text-sm text-muted-foreground">
-      Learn how Coda simplifies Solana development.
-    </p>
-  </a>
-
-  <a
+  />
+  <Card 
+    title="Configuration" 
+    description="Advanced configuration options for customizing code generation to your specific needs."
     href="/docs/configuration"
-    className="group relative rounded-lg border border-border bg-card p-6 transition-colors hover:bg-accent"
-  >
-    <h3 className="font-semibold text-lg mb-2">Configuration</h3>
-    <p className="text-sm text-muted-foreground">
-      Advanced configuration options for customizing code generation to your
-      specific needs.
-    </p>
-  </a>
-</div>
+  />
+</Cards>
