+++ apps/docs/scripts/generate-api-docs.ts
