+++ apps/docs/src/app/layout.tsx
@@ -1,14 +1,33 @@
-import '@/app/global.css';
-import { RootProvider } from 'fumadocs-ui/provider';
-import { Inter } from 'next/font/google';
+import "@/app/global.css";
+import { RootProvider } from "fumadocs-ui/provider";
+import type { Metadata } from "next";
+import { Geist, Geist_Mono } from "next/font/google";
 
-const inter = Inter({
-  subsets: ['latin'],
+export const metadata: Metadata = {
+  title: {
+    template: "%s | Coda | Ian Macalinao",
+    default: "Coda | Ian Macalinao",
+  },
+  description: "Automated TypeScript client generation for Solana programs",
+};
+
+const geist = Geist({
+  variable: "--font-sans",
+  subsets: ["latin"],
 });
 
-export default function Layout({ children }: LayoutProps<'/'>) {
+const mono = Geist_Mono({
+  variable: "--font-mono",
+  subsets: ["latin"],
+});
+
+export default function Layout({ children }: LayoutProps<"/">) {
   return (
-    <html lang="en" className={inter.className} suppressHydrationWarning>
+    <html
+      lang="en"
+      className={`${geist.variable} ${mono.variable}`}
+      suppressHydrationWarning
+    >
       <body className="flex flex-col min-h-screen">
         <RootProvider>{children}</RootProvider>
       </body>
