--- docs/.gitignore
@@ -1,3 +1,0 @@
-.next/
-.source/
-out/
