--- docs/eslint.config.mjs
@@ -1,59 +1,0 @@
-import js from "@eslint/js";
-import nextPlugin from "@next/eslint-plugin-next";
-import reactPlugin from "eslint-plugin-react";
-import reactHooksPlugin from "eslint-plugin-react-hooks";
-import typescriptPlugin from "@typescript-eslint/eslint-plugin";
-import typescriptParser from "@typescript-eslint/parser";
-
-export default [
-  {
-    ignores: [
-      "dist",
-      "node_modules",
-      ".next/",
-      ".source/",
-      "out/",
-      "next.config.mjs",
-      "postcss.config.mjs",
-      "tailwind.config.js",
-    ],
-  },
-  js.configs.recommended,
-  {
-    files: ["**/*.{ts,tsx,js,jsx,mjs}"],
-    languageOptions: {
-      parser: typescriptParser,
-      parserOptions: {
-        ecmaVersion: "latest",
-        sourceType: "module",
-        ecmaFeatures: {
-          jsx: true,
-        },
-      },
-    },
-    plugins: {
-      "@typescript-eslint": typescriptPlugin,
-      react: reactPlugin,
-      "react-hooks": reactHooksPlugin,
-      "@next/next": nextPlugin,
-    },
-    rules: {
-      ...typescriptPlugin.configs.recommended.rules,
-      ...reactPlugin.configs.recommended.rules,
-      ...reactHooksPlugin.configs.recommended.rules,
-      ...nextPlugin.configs.recommended.rules,
-      "no-console": "off",
-      "react/react-in-jsx-scope": "off",
-      "react/prop-types": "off",
-      "@typescript-eslint/no-unused-vars": [
-        "warn",
-        { argsIgnorePattern: "^_" },
-      ],
-    },
-    settings: {
-      react: {
-        version: "detect",
-      },
-    },
-  },
-];
