--- docs/mdx-components.tsx
@@ -1,16 +1,0 @@
-import defaultMdxComponents from "fumadocs-ui/mdx";
-import { Accordion, Accordions } from "fumadocs-ui/components/accordion";
-import * as FilesComponents from "fumadocs-ui/components/files";
-import * as TabsComponents from "fumadocs-ui/components/tabs";
-import type { MDXComponents } from "mdx/types";
-
-export function useMDXComponents(components: MDXComponents = {}): MDXComponents {
-  return {
-    ...defaultMdxComponents,
-    ...TabsComponents,
-    ...FilesComponents,
-    Accordion,
-    Accordions,
-    ...components,
-  };
-}
