--- docs/next.config.mjs
@@ -1,14 +1,0 @@
-import { createMDX } from "fumadocs-mdx/next";
-
-const withMDX = createMDX();
-
-/** @type {import('next').NextConfig} */
-const config = {
-  output: "export",
-  reactStrictMode: true,
-  experimental: {
-    webpackMemoryOptimizations: true,
-  },
-};
-
-export default withMDX(config);
