--- docs/package.json
@@ -1,40 +1,0 @@
-{
-  "name": "@macalinao/coda-docs",
-  "version": "0.0.1",
-  "private": true,
-  "type": "module",
-  "scripts": {
-    "build": "next build",
-    "dev": "next dev",
-    "start": "next start",
-    "postinstall": "fumadocs-mdx",
-    "clean": "rm -rf .next dist content/docs/api/clients",
-    "lint": "next lint",
-    "generate:api": "bun scripts/generate-api-docs.ts"
-  },
-  "dependencies": {
-    "@shikijs/rehype": "^3.11.0",
-    "@shikijs/transformers": "^3.11.0",
-    "fumadocs-core": "^15.6.12",
-    "fumadocs-mdx": "^11.7.5",
-    "fumadocs-ui": "^15.6.12",
-    "lucide-react": "^0.540.0",
-    "next": "^15.5.0",
-    "react": "^19.1.1",
-    "react-dom": "^19.1.1",
-    "shiki": "^3.11.0"
-  },
-  "devDependencies": {
-    "@tailwindcss/postcss": "^4.1.12",
-    "@types/mdx": "^2.0.13",
-    "@types/node": "^24.3.0",
-    "@types/react": "^19.1.10",
-    "@types/react-dom": "^19.1.7",
-    "eslint": "^8.57.1",
-    "eslint-config-next": "^15.5.0",
-    "postcss": "^8.5.6",
-    "tailwindcss": "^4.1.12",
-    "typescript": "^5.9.2",
-    "yaml": "^2.8.1"
-  }
-}
