--- docs/postcss.config.mjs
@@ -1,8 +1,0 @@
-/** @type {import('postcss').ProcessOptions} */
-const config = {
-  plugins: {
-    "@tailwindcss/postcss": {},
-  },
-};
-
-export default config;
