--- docs/source.config.ts
@@ -1,21 +1,0 @@
-import { rehypeCodeDefaultOptions } from "fumadocs-core/mdx-plugins";
-import { defineConfig, defineDocs } from "fumadocs-mdx/config";
-
-export const docs = defineDocs({
-  dir: "content/docs",
-});
-
-export default defineConfig({
-  lastModifiedTime: "git",
-  mdxOptions: {
-    rehypeCodeOptions: {
-      themes: {
-        light: "catppuccin-latte",
-        dark: "catppuccin-mocha",
-      },
-      inline: "tailing-curly-colon",
-      defaultLanguage: "typescript",
-      transformers: [...(rehypeCodeDefaultOptions.transformers ?? [])],
-    },
-  },
-});
