--- docs/src/app/docs/layout.tsx
@@ -1,12 +1,0 @@
-import { DocsLayout } from "fumadocs-ui/layouts/docs";
-import type { ReactNode } from "react";
-import { source } from "@/lib/source";
-import { baseOptions } from "@/app/layout.config";
-
-export default function Layout({ children }: { children: ReactNode }) {
-  return (
-    <DocsLayout {...baseOptions} tree={source.pageTree}>
-      {children}
-    </DocsLayout>
-  );
-}
