--- docs/src/app/layout.config.tsx
@@ -1,15 +1,0 @@
-import type { BaseLayoutProps } from "fumadocs-ui/layouts/shared";
-
-export const baseOptions: BaseLayoutProps = {
-  nav: {
-    title: "Coda",
-  },
-  githubUrl: "https://github.com/macalinao/coda",
-  links: [
-    {
-      text: "Documentation",
-      url: "/docs",
-      active: "nested-url",
-    },
-  ],
-};
