--- docs/src/lib/source.ts
@@ -1,7 +1,0 @@
-import { loader } from "fumadocs-core/source";
-import { docs } from "@/.source";
-
-export const source = loader({
-  baseUrl: "/docs",
-  source: docs.toFumadocsSource(),
-});
